+++ app/src/main/AndroidManifest.xml
@@ -27,6 +27,8 @@
         <activity android:name=".activity.SetLocationActivity"
             android:theme="@style/Theme.AppCompat.Light.Dialog"/>
 
+        <activity android:name=".activity.ScheduleInfoActivity"
+            android:theme="@style/Theme.AppCompat.Light.Dialog"/>
     </application>
 
 </manifest>
