+++ app/src/main/res/layout/dialog_calendar.xml
@@ -1,9 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
+<android.support.constraint.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@android:color/white"
+    android:layout_weight="1"
     android:orientation="vertical">
 
     <LinearLayout
@@ -17,41 +20,22 @@
             layout="@layout/current_date"
             android:layout_width="match_parent"
             android:layout_height="0dp"
-            android:layout_weight="1"
+            android:layout_weight="0.1"
             tools:ignore="NestedWeights" />
-<!--
+
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="0dp"
-            android:layout_weight="1"
-            android:orientation="vertical">
-        <TextView
-            android:id="@+id/tv_date"
-            android:layout_width="match_parent"
-            android:layout_height="30dp"
-            android:background="@drawable/rounded_lightgreen"
-            android:gravity="center"
-            android:textColor="@color/defaultColor"
-            android:textSize="16dp"/>
-
-        <Space
-            android:layout_width="match_parent"
-            android:layout_height="5dp" />
-
-        <View
-            android:layout_width="match_parent"
-            android:layout_height="1dp"
-            android:background="@color/defaultColor"/>
-        </LinearLayout>-->
-
-        <FrameLayout
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="5">
+            android:layout_weight="0.8"
+            android:focusable="true"
+            android:focusableInTouchMode="true"
+            android:orientation="vertical"
+            app:layout_behavior="@string/appbar_scrolling_view_behavior">
 
             <LinearLayout
                 android:layout_width="match_parent"
-                android:layout_height="match_parent"
+                android:layout_height="0dp"
+                android:layout_weight="0.7"
                 android:gravity="bottom"
                 android:orientation="horizontal">
 
@@ -69,22 +53,22 @@
                     android:gravity="center_vertical"
                     android:textSize="25dp" />
             </LinearLayout>
-        </FrameLayout>
-    </LinearLayout>
 
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="40dp"
-        android:background="@color/colorGreen">
-
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="0dp"
+                android:layout_weight="0.1">
         <Button
             android:id="@+id/btn_cancel"
             android:layout_width="match_parent"
             android:layout_height="40dp"
             android:layout_gravity="center_vertical"
             android:text="@string/exit"
-            android:background="@drawable/background_transparent"
+            android:background="@color/colorGreen"
             android:textSize="16dp"
             android:textColor="@color/colorWhite"/>
+            </LinearLayout>
     </LinearLayout>
-</LinearLayout>
+    </LinearLayout>
+
+</android.support.constraint.ConstraintLayout>
